Move get valid plugin info to shared space
